+++ apps/api/src/models/owners/owners.resolver.ts
@@ -1,19 +1,44 @@
-import { Resolver, Query, Mutation, Args } from '@nestjs/graphql'
+import {
+  Resolver,
+  Query,
+  Mutation,
+  Args,
+  ResolveField,
+  Parent,
+} from '@nestjs/graphql'
 import { OwnersService } from './owners.service'
 import { Owner } from './entities/owner.entity'
 import { FindManyOwnerArgs, FindUniqueOwnerArgs } from './dto/find.args'
 import { CreateOwnerInput } from './dto/create-owner.input'
 import { UpdateOwnerInput } from './dto/update-owner.input'
+import { Billboard } from '../billboards/entities/billboard.entity'
+import { BillboardStatus } from '@prisma/client'
+import { PrismaService } from 'src/common/prisma/prisma.service'
+import {
+  AllowAuthenticated,
+  GetUser,
+} from 'src/common/decorators/auth/auth.decorator'
+import { GetUserType } from '@billboards-org/types'
+import { checkRowLevelPermission } from 'src/common/guards'
 
 @Resolver(() => Owner)
 export class OwnersResolver {
-  constructor(private readonly ownersService: OwnersService) {}
+  constructor(
+    private readonly ownersService: OwnersService,
+    private readonly prisma: PrismaService,
+  ) {}
 
+  @AllowAuthenticated()
   @Mutation(() => Owner)
-  createOwner(@Args('createOwnerInput') args: CreateOwnerInput) {
+  createOwner(
+    @Args('createOwnerInput') args: CreateOwnerInput,
+    @GetUser() user: GetUserType,
+  ) {
+    checkRowLevelPermission(user, args.uid)
     return this.ownersService.create(args)
   }
 
+  @AllowAuthenticated('admin')
   @Query(() => [Owner], { name: 'owners' })
   findAll(@Args() args: FindManyOwnerArgs) {
     return this.ownersService.findAll(args)
@@ -24,13 +49,27 @@ export class OwnersResolver {
     return this.ownersService.findOne(args)
   }
 
+  @AllowAuthenticated()
   @Mutation(() => Owner)
-  updateOwner(@Args('updateOwnerInput') args: UpdateOwnerInput) {
+  updateOwner(
+    @Args('updateOwnerInput') args: UpdateOwnerInput,
+    @GetUser() user: GetUserType,
+  ) {
+    checkRowLevelPermission(user, args.uid)
     return this.ownersService.update(args)
   }
 
+  @AllowAuthenticated()
   @Mutation(() => Owner)
-  removeOwner(@Args() args: FindUniqueOwnerArgs) {
+  removeOwner(@Args() args: FindUniqueOwnerArgs, @GetUser() user: GetUserType) {
+    checkRowLevelPermission(user, args.where.uid)
     return this.ownersService.remove(args)
+  }
+
+  @ResolveField(() => [Billboard])
+  billboards(@Parent() billboardStatus: BillboardStatus) {
+    return this.prisma.billboard.findUnique({
+      where: { id: billboardStatus.billboardId },
+    })
   }
 }
